backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-ratelimit==4.1.0
django-filter==23.5
python-decouple==3.8
Pillow==10.1.0
mysqlclient==2.2.0
gunicorn==21.2.0

# ML/AI
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2

# Image processing
pillow==10.1.0

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
